A cast to a byte, storing through a pointer to a byte
Four of them - `*global_ptr(hit->cell_ptr) = (uint8_t)to`, `*name++ =
(uint8_t)c`, `*dst = (uint8_t)out` and plot_pixel's, whose expression is
provably under 256 anyway. C truncates on the store, so each said the
same thing twice.

The object files are byte-identical to the commit before this one, every
section and every symbol, so this is checked by comparing what was built.

Rule 10 could not see any of them, and that is the part worth fixing: its
left-hand side allowed a name, a field and a subscript, but neither a `*`
nor a call. A store through a pointer is a target like any other, and its
width comes from one of two places - the accessor's return type for
`*global_ptr(...)`, and the local's own declaration for `*dst`.

That second one needed two passes. The declaration map wants `type name`
and cannot see the `*` in `uint8_t *dst`, so the first widening still
missed two of the four; pointer locals now carry the width of what a
store through them lands in, with the same rule as the other map that a
name declared at two widths is unresolvable rather than guessed.

Checked by putting all four casts back and watching the rule report them,
rather than by observing that it reports nothing - which it also did
before it worked.

+++ reconstruct/src/game.c
@@ -2357,7 +2357,7 @@ static entity_t *brick_entity(hit_t *hit, uint16_t handler_fn,
  * every other brick is drawn from. */
 static void brick_degrade(hit_t *hit, uint8_t from, uint8_t to)
 {
-    *global_ptr(hit->cell_ptr) = (uint8_t)to;
+    *global_ptr(hit->cell_ptr) = to;
     uint8_t x = hit->centre.x, y = hit->centre.y;
     xor_sprite_16x7(x, y, global_ptr(global.cell_bitmap.plain_ptr[from]));   /* rub out */
     xor_sprite_16x7(x, y, global_ptr(global.cell_bitmap.plain_ptr[to]));     /* and draw */
@@ -4806,7 +4806,7 @@ int32_t name_field(uint16_t di, uint8_t *abort)
                   (c >= '0' && c <= '9') || (c >= 'A' && c <= 'Z'));
         if (!ok || len == 12)
             continue;
-        *name++ = (uint8_t)c;
+        *name++ = c;
         len++;
         draw_char((char)c, di);
         draw_cursor(di);
@@ -5326,7 +5326,7 @@ void plot_pixel(uint16_t x, uint16_t y, uint8_t colour)
     uint16_t di = cga_at(x, y);
     uint8_t shift = 6 - (x & 3) * 2;
     uint8_t *p = &g_vram[di & (CGA_SIZE - 1)];
-    *p = (uint8_t)((*p & ~(3u << shift)) | ((colour & 3) << shift));
+    *p = (*p & ~(3u << shift)) | ((colour & 3) << shift);
 }
 
 /* The same with bit 7 of AL set: XOR rather than replace. */
@@ -6769,7 +6769,7 @@ void screen_end_of_game(void)
                 if ((old & mask) == 0)
                     out += add & mask;
             }
-            *dst = (uint8_t)out;
+            *dst = out;
         }
 
         /* One band on screen from the saved copy, then the merged block. */
